requirements.txt
hidapi>=0.7.99.post21
PyCRC>=1.21
pytest-pylint
